+++ Client/N3Base/N3Snd3dObj.h
@@ -19,7 +19,7 @@ protected:
 	
 public:
 	void SetConeOrientation(D3DVECTOR* pDir);
-	void SetConeOutSizeVolume(LONG vol);
+	void SetConeOutSizeVolume(int32_t vol);
 	void SetMaxDistance(D3DVALUE max);
 	void SetMinDistance(D3DVALUE min);	
 	void SetPos(const D3DVECTOR* pvPos);
